adapter: Remove upstream schema validation
The code to validate ReadySet-returned schemas against the upstream was
never completed or used in the real world, and we've been paying the
cost for maintaining it. It's unlikely that we'll ever want to actually
do this, so instead let's just remove it to save on the maintenance
cost.

There might have been *something* I missed during this removal -
basically I just removed the configuration flags and chased "identifier
does not exists" and "dead code" warnings until everything looked happy.
If anything is found that was only needed for this later, we can remove
it down the road.

Fixes: ENG-2954
Change-Id: I53ad4471b082473b1f71dd510647f5e5358411b6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4806
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>