readyset: Make migration task upstream conn fault tolerant
When running with async migrations, the migration background task needs
to load the server-configured schema search path from the upstream db
on startup. Previously, this process was not fault tolerant, and would
cause a panic triggering a full-process abort if the connection failed.
This commit changes it to sleep a second and automatically retry if it
encounters an error instead.

Fixes: ENG-2835
Change-Id: I8d464a55998029704ec79e9a031d8d30f006221c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4807
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>