psql-srv: Discard errors during extended query flow
The postgresql protocol message flow documentation[0] specifies that
between the beginning of an "extended query" flow and a "sync" message,
all but the first error response should be discarded. Some client
libraries take advantage of this behavior to unconditionally send
multiple packets for an extended query without checking the response for
an error each time (prisma-migrate, for example, sends a Parse packet
followed by a Describe packet unconditionally, without checking the
Parse packet for errors first). Previously, we were behaving incorrectly
here - we'd return *all* errors for all packets within an extended
query. This meant that an error encountered in a subsequent packet (in
this case, the fact that the Describe referenced a statement that didn't
exist, since the Parse failed) would get associated with the *wrong
request* (in our case, a request to create a prisma shadow database).

Instead, this commit:

1. Makes sure we enter the Extended state when handling the Parse
   message, since it seems both empirically and from carefully reading
   the docs that that is supposed to put the server in an Extended state
2. Discards all messages after the first error until we see a Sync
   packet, by just matching on State::Error in on_request and returning
   Response::Empty for all non-Sync messages.

There were a few TODOs scattered around noting that we should be doing
this, which have also been removed.

Note that this is not on its own sufficient to make Prisma migrations
work when proxied through ReadySet, as we still ignore the database name
provided upon connection.

[0]: https://www.postgresql.org/docs/current/protocol-flow.html

Fixes: ENG-2566
Change-Id: Ic894e46a25930cdb500491edf8bbc4b0ef991322
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4811
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>