database-utils: Reorganize database-utils
This commit creates a new `connection` module to house
`DatabaseConnection` and all of the related types. This sets the stage
for future commits, which will add more connection-specific types to
the module (e.g. a type to handle transactions). This commit also
updates the trait bounds on the query parameter types for
`DatabaseConnection`'s methods to improve clarity and flexibility.
Specifically, the traits bounds have been updated to allow callers to
pass `DfValue`s as parameters directly. Finally, this commit also
removes a type parameter on the `DatabaseError` type, replacing it with
a trait object. The type parameter was only used in one of the variants,
so in practice, the never type was passed more often than not. Replacing
the generic member with a trait object allows us to capture
error-specific information in a way that is less verbose and more
flexible.

Change-Id: Ia851eec10478cadf492e9814c97fb65a44a8c44c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4843
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>