database-utils: Add `QueryResults`
This commit adds a `QueryResults` type to support query results that may
arise from either a MySQL- or PostgreSQL-backed `DatabaseConnection`.
Having `DatabaseConnection` methods return a database-agnostic type to
house raw results simplifies and makes concrete the return types of the
query methods and allows callers of these methods to decide how to
handle results in a more flexible way, as opposed the previous design,
which forced callers to pass a type parameter to which the query results
would always be converted, even in situations where they didn't need to
be converted at all.

Change-Id: Ibd78fafd84c3af209a2f5743318800eaac4b7885
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4844
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>