helm: Chart 2.0-ish
This is an overhaul and update of the original Helm chart. The changes
herein steer it towards best practices and bring the Consul dependencies
up to date (1.15.1 as of this commit (chart version 1.1.1)).

We have a good start to start building on much needed changes like
adding support for expanding PVCs, providing ways to expose the
ReadySet HTTP endpoints for monitoring, and general user experience
improvements like a useful NOTES.txt for when the chart install
completes.

- Rename template files according to <function>-<resource>.yaml as this
  is close to the noted best practice
- Update annotations on AWS LB Controller for readyset-adapter to work
  with the updated version of the AWS LB Controller users will install.
- Remove scripts/ directory
- Add a definition for "readyset.name"
- Remove references to .global.*
- Simplified names of resources and references to them preventing
  issues like repeated chart names
- Set some low, but reasonable defaults for EKS and GKE autopilot
- Remove env vars set to the already default value
- Remove duplicative command definitions for server and adapter

Change-Id: I5d406586ad808dd06f021b4b6c9aed8135ed678c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4816
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Jason Camp <jason@readyset.io>