logictest: Drop the database in between generate runs
I'm seeing some strange, inconsistent behavior with queries that the
fuzzer is running, and I'm not sure what's causing it - I have
a *hypothesis* that there's an issue with some state getting left behind
between runs, so for now let's try just dropping the whole database in
between runs instead of dropping each table by hand. Even if this
doesn't fix the issue, I'm betting it'll be faster.

Change-Id: I0bc4e1e0a2d516ab780ecaee8fefb25440c5f7ed
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4817
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>