query-generator: Expose dialect in generate opts
Thread the dialect parameter specifically used to display queries up
through the various options structs used to parametrize query
generation, both within query-generator itself and in logictest. For the
logictest commands in particular, the dialect is not specified
explicitly by the user, but is inferred from the database type of the
database URL that's being compared against.

Fixes: ENG-2498
Change-Id: Idbbe7bae5cc0a3092edcb00132dbdf83d559fcc8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4820
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>