psql-srv: Pass through *all* error fields from upstream
Pass through the full set of upstream error fields when converting
upstream errors to error packets to send back to the client, not just
the ones that're guaranteed to be always present. This creates a nicer
error-reporting experience in the psql client when errors are
encountered, and also *maybe* more client libraries will be tightly
coupled to specific error fields (though I hope not, since none of these
are guaranteed to be present).

Change-Id: I4e5be3780472e5337e12b1a12b908e8989b33378
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4826
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>