mir: DISTINCT nodes don't need __distinct_count as input
Without this, DISTINCT nodes would think they needed to pull
their (output!) `__distinct_count` column from their parents, and if
they *found* that column (eg in a DISTINCT node made for a distinct
aggregate somewhere earlier in the query) they'd pull it in a way that
wouldn't always project the column properly (eg if it went through a
JoinAggregates node). DISTINCT nodes only need their group_by as input,
which resolves this issue.

Fixes: ENG-2981
Change-Id: I3cb6418fa0c97677a0ce92868c18dc12e7db9516
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4832
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>