benchmarks: Reduce heap allocations
Previously, `DatabaseStatement` had a `Str` variant that held an owned
`String` whose sole purpose was to allow callers of
`DatabaseConnection::execute` to bypass having to call
`DatabaseConnection::execute_str` directly. This resulted in unncessary
conversions from `&str` to `String` when `&str` was passed to
`DatabaseConnection::execute` due to the `DatabaseStatement:
From<&str>` implementation. This commit removes the
`DatabaseStatement::Str` variant and generalizes the method signature of
`QueryableConnection::execute` to support passing anything that
implements the new `DatabaseStatementLike` trait (including
`DatabaseStatement`s and `&str`s).

In `QueryBenchmark`, we were previously creating a single prepared
statement to be shared across the connections in each of the benchmark
threads, which isn't how prepared statements work. It wasn't causing
issues because we were just passing the query as a String to
`DatabaseConnection::execute`, which caused the underlying library to
prepare the query automatically for us for each of the connections. Now
that `PreparedStatement::generate_query` returns the actual prepared
statement instead of the raw query string, this commit prepares a
statement in each benchmark thread.

Finally, `Hash`, `PartialEq`, and `Eq` are implemented for
`DatabaseStatement` to enable us to use `DatabaseStatement` directly in
the `benchmarks` crate rather than using the raw query string.

Change-Id: I770044558b37a055767db9eb8e6f78c17b22ce15
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4847
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>