query-generator: Generate type-appropriate min/max for BETWEEN
Use `FilterRHS` within the op for both the min and max values rather
than hard-coding integer literals, so that the types of the min and max
match the type of the column we're comparing against. This matters for
postgresql, which returns errors if we compare against a mismatched data
type, but also gets us BETWEEN comparisons against columns for free,
which is nice in terms of coverage.

Refs: ENG-2988
Change-Id: Ifc45cb35afe616186c8bf5e6a011d7aab3a90a5c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4839
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>