Return correct CommandComplete for transactions
Previously we returned an empty CommandComplete message when handling
transaction messages. This CL delegates transaction commands to the
tokio-postgres simple_query() to make use of the work we did to pass
CommandComplete tags through to the client.

Since BEGIN and START TRANSACTION are distinct CommandComplete tags,
this CL also extends nom_sql::StartTransactionStatement to differentiate
between the two.

Fixes: ENG-2950
Change-Id: Ib165214973a2ea70c83d87aab3002af12349f760
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4841
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>