Revert "query-generator: Manually impl Arbitrary for QueryOperation"
This reverts commit cbdce9b7e63ea95a0223f43b83a6c50451f05708.

Turns out the actual issue here wasn't the `#[weight(u32::from(bool))]`
trick (I should've known, since that actually *panics* if the weight is
0!), it was that we were never actually setting in_subquery to false,
because we weren't actually using the Arbitrary impl for Subquery! I'm
going to clean this up so everything works properly, but in the
meantime I'm just reverting this commit directly so we get a clean
revert.

Refs: ENG-2988
Change-Id: Icf9525d02ccb073b3c26144d2bebcd2ef5207ace
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4849
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>