query-generator: Fix skipping of parameters in subqueries
The previous approach of passing in an `in_subquery` boolean from the
Arbitrary impl for Subquery to the parameters for QueryOperation wasn't
working, because we were never actually using the Arbitrary impl for
Subquery itself (since we want to use prop_recursive in QuerySeed to
avoid blowing up the stack). It in fact *wouldn't* have ever worked,
since proptest panics if the weight used for a variant in a Union is
ever 0.

Instead, to get this working in the easiest (though maybe not most
optimal) way possible, we just filter out any operations that aren't
supported in subqueries when constructing subqueries during generation
of QuerySeeds.

Refs: ENG-2988
Change-Id: I810dc6bf93af521e1518efe133570d7dd1508cb1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4854
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>