logictest: Install the CITEXT extension if necessary
In order to create tables with columns that have the CITEXT type, we
need to explicitly install the `CITEXT` extension in postgres via a
`CREATE EXTENSION` statement. I iterated a bit through the best way to
do this, and eventually landed on doing it *as part of the test script*,
as an extra statement record before the actual test. This has the nice
property that the script can be run verbatim on a fresh, unmodified
postgresql database.

Change-Id: I7cac3da463e974057fcde7b24d2fb4448cef4307
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4858
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>