mysql-srv: Decode length-4 FIELD_TYPE_DATETIME values
Empirically, it looks like some mysql client libraries will send only a
year, month, and date when sending FIELD_TYPE_DATETIME where the hour,
minute, and second are all 0 - even though this does not seem to appear
in the documentation - and the mysql server itself will accept this.
This commit updates the conversion from mysql_srv::Value to
NaiveDateTime to handle this.

Fixes: ENG-2997
Change-Id: I03ec0916ff6bf0da9898997f2b46211ae2db5271
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4866
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>