data-generator: Don't try to generate invalid times
In the branch of unique_value_of_type for TimestampTz, we were just
passing the unique index to the `minute` field, which would mean that
any unique idx above 59 would cause chrono to panic due to us trying to
initialize an invalid time. Instead, this just *adds* the idx in minutes
to a base time.

Change-Id: Iccb2db4b4e23e96764a805f156275520715b9c6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4872
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>