dataflow: Simplify arguments to Join::new
Currently the way we specify the column configuration for a Join
operator is a little convoluted - we're passed a vec of  an enum
JoinSource which has variants for left, right, and "both", and the join
key is extracted from the Both elements. This is especially silly since
we actually store the join key and columns to project separately in the
MIR node, so there's really no point in having this strange intermediary
representation - not to mention that the complexity here has been (for a
very long time) hiding a bug (ENG-3013).

This commit starts the process of simplifying the mess, by getting rid
of JoinSource and just passing the on columns and emit columns in the
format we want to store them in Join directly. The way this is all
constructed in mir_to_flow is still a little convoluted and still
contains ENG-3013, but that'll be fixed next.

Refs: ENG-3013
Change-Id: I7c3feafe793e37409741102c57fc1407b64b2541
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4885
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>