mir-to-flow: Construct join key directly from `on`
Rather than constructing the join key for a Join node by iterating over
the projected columns, construct it directly from the `on` columns taken
from the MIR Join node. This is not only much simpler and more direct,
but also fixes a miscompilation bug where we'd skip all but one column
in a composite join key where the left column was repeated (see the
logictest, which was failing previously, for an example).

Fixes: ENG-3013
Change-Id: I01114a801f301f661ac81443cecde3bfa10f9358
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4886
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>