readyset-mysql: Fix issues with foreign key vertical tests
The new stateful-proptest version of the dataflow vertical tests was
crashing during test generation in a couple of different places. The
issue was that we would sometimes try to generate inserts and updates
for tables with foreign keys when we didn't have any rows in the foreign
table to select a key from.

For the old version of the vertical test, this kind of issue would
usually be caught by a filter, and since all the generators were being
squished together into a `prop_oneof!` call, proptest would just retry
and pick a different operation. With stateful-proptest though, we
randomly pick a single generator and then generate directly from that,
so retries just try again with the same generator and hit the same
filter repeatedly until proptest throws an error due to the max number
of filter rejections is hit.

There is more weirdness and subtlty beyond this in the way that the old
vs. new vertical test generation differs, so ultimately I decided to
just rework some of the generation code a bit. We now calculate which
tables have foreign keys on which other tables, and look directly at the
rows in the other tables to determine whether we can generate rows that
contain valid foreign keys for insert and update operations.

Refs: ENG-2839
Change-Id: I6a8f09329a38a05037894f127bd61a2adc1ad667
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4898
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>