dataflow: Don't create spurious parent indexes in aggs
In 55d6dbaf6 (Teach Extremum how to recover from a deleted extremum
value, 2021-06-21), we made it so that if we delete the single extreme
value from an Extremum aggregate we can do a lookup into the parent node
to repopulate the state of the node. To support that, we create a lookup
obligation on the group_by cols in the parent node - but we do that
for *all* Grouped operators, not just Extremum, even though only
Extremum needs it! This has the effect of materializing *all the rows*
in the table in that parent node, even though we'll never need to do a
lookup into that materialization!

This fixes that, by only actually requesting that parent indexing
obligation if the underlying grouped operator needs it (which is only
true for Extremum).

Change-Id: I4813454c503c09d817a5604447a7e9848e4ac2e9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4900
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>