mir: Don't make bogo group keys
Now that nodes can be indexed on an empty set of columns, we no longer
need to create "bogo" group keys for ungrouped aggregates in queries -
instead, we can just let the group_by cols for those aggregates be
empty! This has a number of advantages: it simplifies compilation, it
saves on an extra node, it might save on some extra materializations,
and we might have never needed to make a bogo-group anyway (eg if we
were going to pull something like a ViewKey or dependent filter through
the aggregate, which'd add to the group_by cols anyway).

Change-Id: I094fd9771b9333ed578bfce12081684781b59d37
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4902
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>