readyset-mysql: Implement preconditions for vertical test ops
This fixes issues that could occur during shrinking if, say, an insert
is shrunk out and then we try to update or delete a row that no longer
exists.

In theory this shouldn't necessarily cause any differing results between
ReadySet and MySQL, but we deliberately don't generate test cases like
this, so producing cases where this can happen during shrinking seems
like a bad idea. Additionally, it can cause our `next_state` code to
crash, since it's written under the assumption that we won't generate an
update or a delete for a row that doesn't exist. Rather than make
`next_state` tolerant to these situations, I think it makes more sense
to use preconditions to enforce the same restrictions during shrinking
that we apply during initial test generation.

Fixes: ENG-3027
Change-Id: I20cdbfbd28098af0b0dda93edfd866c49267c6aa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4905
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>