replicators: Don't panic if our snapshotting estimate is off
Now that as of 950a4e76f (replicators: Use a row estimate for
snapshotting progress, 2023-05-09) we're using an *estimate* for row
counts to report snapshotting progress rather than an exact count, it
can be the case that we snapshot more rows than we originally thought.
This commit fixes an overflow panic in the "amount of time remaining"
estimate calculation if that's the case, by changing a duration
subtraction to use `saturating_sub` instead.

Fixes: ENG-3031
Change-Id: I1a37fcdea559d300e6aa90f2cf061a16b8402350
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4906
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI