domain: Log progress during persistent state init
If initializing an individual persistent state takes longer than a
certain time (I'm starting with 15 seconds), log a message saying that
it's still happening using an exponential backoff, to give the user
feedback that that is going on in the background. Then, once it's open,
if we logged at least once (eg if we passed the 15 second mark), log to
say that it's open. All of this is done with a new
`report_progress_with` helper function, which is intended to generalize
this kind of pattern of logging periodically while a long-running task
is happening. Hopefully this strikes the balance of giving users
feedback if things take a while, but not being too overly verbose in the
"happy path" if things are quick

Change-Id: I66146baae731209eb4f6436988282761df7dcfbb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4907
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>