stateful-proptest: Work around issue with proptest source_file field
The `proptest!` macros hardcode the `source_file` field of
`ProptestConfig` to `file!()`, regardless of what is passed in. This
seems like it might be a bug, since `source_file` is `Option`, so if
it's not the default value of `None`, it would seem like the macro could
be smart enough to leave it alone.

However, until such time as this might be addressed upstream, we can
work around this by manually creating a `TestRunner` and replicating
some of the code that would otherwise by created by the macro.

Refs: ENG-3033
Change-Id: I00c8b42f5064818c294144fb56586f9c332db3f0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4909
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>