mir/rewrite: Don't try to lift filters if their cols stop resolving
In filters_to_join_keys, if we're looking past a point where the columns
of a join-key-candidate filter actually resolve through the parent, we
should just stop trying to turn that filter into a join key rather than
doing what we previously did and returning an error. This fixes eg
filters that compare two columns from the same table, which can't be
turned into join keys.

Change-Id: I3f8246d69cf96b1ccdb7645caa7125c364522a74
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4911
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>