persistent-state: run compaction in the background
Rather than running the post-snapshotting manual compaction in the
foreground, blocking both the domain and the table RPC, run compaction
for *each* column family in a background thread, using the fact that we
now have a shared reference to an RwLock around the database (and the
fact that manual compaction doesn't need exclusive access to the db
handle!). Over on the replicators side, we use a new
`"all_tables_compacted"` RPC to periodically poll each base-table domain
to ask if it's finished compaction - but only once we've started
compaction for all snapshotted tables.

To allow compaction to be *canceled* if the user shuts down readyset,
this takes advantage of the new "cancelation flag" I added to
CompactOptions in the rocksdb C api and rust-rocksdb, setting it to true
if the compaction thread handle (which is part of PersistentState) is
dropped.

Change-Id: I1e51dc5def103698fa01b5f05911d03c514cf4f2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4937
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>