dataflow: Make DfState auto-impl Send + Sync
This commit contains changes that allow `DfState` to auto impl Send +
Sync. The blockers for those being implemented were two RefCells that
held mutable state for Concat and Aggregation nodes.

Instead of holding them in the Graph itself, we now hold a per-node
mapping of AuxiliaryNodeState at the domain level that is passed in for
processing.

Allowing DfState to be Send + Sync is a prerequisite for making
Leader::external_request async and non-blocking.

Change-Id: I954181f60f92808ff326b1b92b56f431b29bc868
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4945
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>