migrate: Wait for IsReady only as we need to replay
Previously, we'd implicitly wait for nodes to be ready as part of
sending the Ready message. This loses on an opportunity for parallelism,
since we only really *need* for nodes to be ready if we're actually
going to do a full replay from them - otherwise we can just kick off the
readying of the various nodes in the background, and wait for anything
that we're not replaying from as part of the same migration to be ready
once everything's running. This does this by no longer *implicitly*
sending IsReady after we send Ready, and instead threading around a
`&mut non_ready_nodes: HashSet<NodeIndex>`, and adding an *explicit*
IsReady message for nodes we haven't already waited to be ready either
before we're about to start a replay from them, or at the end of all
the rest of the materialization setup messages.

A previous version of this, 3111dcaa1 (migrate: Ready all nodes in
parallel, then wait for ready, 2023-05-17), waited to send *all* IsReady
messages until the end, which worked in tests initially since in tests
persistent state initializes quite quickly, but broke in real-world
scenarios since we'd try to start a replay from a non-ready base node.

Change-Id: I2a28f43a1bacc241f9cb14cba81ff76b0f5e55c6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4950
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>