Domain: Implement random evictions by Tag
Implements random single key evictions per Tag in Domain. Creates a new
EvictRequest::Random { Tag }.

This request is handled by Domain::handle_eviction(). The Domain evicts
a single key from the destination of the given Tag, and if the node was
not a reader, the Domain calls trigger_downstream_evictions() for the
evicted key, target node, and index returned from state::evict_random().

Any partially materialized index can be evicted from in this way, as
each index must be the destination of some replay path.

Change-Id: I04e4fe3735578111f9ac3f58fba06cf511e12c9f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4961
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>