readyset-server: Make external_request an async fn
Now that DfState is Send + Sync, we can make external_request and async
fn.

I'm not entirely sure if it's a compiler limitation or bug, but the way
we constructed an iterator in `node_sizes()` was blocking compilation,
so I added a small allocation to this function to workaround the liftime
errors. Since this is called in non-performance critical code (tests and
graphviz RPCs), this is acceptable.

Change-Id: Iccf10a289d7b9d90c674426c5124a5cad79347cb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4952
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>