server: Add test for multiple simultaneous migrations
As we make more and more things happen in parallel, and especially with
the way we're cloning DfState and then writing back to commit it, it's
important that we make sure that multiple simultaneous migrations don't
step on each others' toes. I had worried that this was *already* broken,
so I wrote this test to check - it seems fine, but I'm committing it
anyway to make sure it doesn't regress.

Change-Id: If9bb3fbc904db1a243e319fa2d4b5cde59450f3a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4985
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>