Controller: Implement /evict_random RPC
Implements an /evict_random RPC in the controller, which selects a
random materialized index in the graph, then delegates the eviction to
the corresponding domain.

Introduces a Materializations::partial_tags() helper which returns a
(NodeIndex, Tag) pair for each partial index in the graph. This set will
include a Tag for every partial index in the graph, since each partial
index must be the destination of a replay path. Multiple Tags may
correspond to the same index for some graphs.

Also introduces a DfState::domain_for_node() helper, which traverses all
domains and nodes in DfState::domain_nodes to find the Domain that owns
a Node.

Change-Id: I464efebb55c327312a5e02c783169e5212f84eee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4996
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>