server: Run dmp.apply() for intitial recovery in background
The process of recovering the state of the graph on startup consists of
two logical steps: a step to load the state from the authority, and a
step to actually send all the domains to workers. The latter step
takes *considerably* longer - up to multiple hours if base tables are
especially large. Currently, we're blocking the entire controller thread
on both steps, which means any controller requests made during that
recovery will time out. Instead, we can run the second phase in the
background (still failing the controller if it returns an error or
panics), and be able to handle most (eg read-only) controller requests
while it's happening.

This means we're committing the dataflow state *twice* during recovery -
once to commit the graph state, and once to commit the updates to the
state made after applying the dmp (such as the addresses of various
domains, etc.)

Fixes: ENG-3057
Change-Id: I83b102c64422461634ac92c941001200c4451e57
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4987
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>