Vertical Dataflow: Add Evict Operator
Adds Operator::Evict to vertical dataflow tests. This operator is
executed by calling the /evict_random Controller RPC.

This operator has no preconditions as calling it when partial state is
empty will result in a no-op.

This operator will evict a random key on each call, so it is not
compatible with test shrinking.

Change-Id: I498bb1a20897f4339e46b2e329e3db0dd10e5059
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4997
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>