dataflow: Simplify arguments to Project node
Previously, the dataflow Project node was strictly limited to always
having a vector of columns to emit verbatim from the parent, followed by
a vector of expressions, followed by a vector of literals. This is more
complex than we need, because all three of these fit perfectly fine into
`Expr`, but more importantly is inflexible, leading us to need to create
multiple project nodes in a row where a simpler `emit: Vec<Expr>` would
do. This commit simplifies the emit argument to Project to just take a
Vec<Expr>, in preparation for making the building of project nodes
simpler and also writing a MIR graph rewrite pass to merge successive
project nodes into one.

Over on the MIR side, we still need to be able to preserve MIR column
aliases, so we introduce a new `ProjectExpr` enum which is either a
`mir::Column` or a `nom_sql::Expr` and an alias.

Change-Id: I28c7e1d6add93466705e92eea2ab637290f2dbd1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5019
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI