mir: Add a rewrite pass to fuse project nodes
I noticed in a flamegraph of the execution of `CREATE CACHE` for a
full-table aggregate of a large `COUNT(*)` that we're spending a
significant amount of time (about 22%, on my machine) executing
`on_input` for `Project` nodes - currently, those queries create 5
project nodes, of which 3 are immediately after another project node,
which is unnecessary now that project nodes can always be fused. To try
to optimize that away a bit, this commit introduces a new MIR rewrite
optimization pass, `fuse_project_nodes`, which fuses MIR Project nodes
that follow each other directly, inlining exprs into column references.
On my machine, this speeds up the full replay for a COUNT of a 10GB
table from around 1m 56s to around 1m 38s.

Change-Id: Iac21a09140bf03799b490bf1555facf40a6d7301
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5020
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>