domain: Use index of target, not destination, for holes
The `Hole`s we store in `Waiting::redo`s are holes in the *targets* of
replay paths, not *destinations* (see [0]) - previously, though, we were
using the destination rather than the target index when iterating
through the keys of a replay path segment to discard keys we weren't
waiting for. This caused us to incorrectly discard a key when we
received a replay along an extended replay path where the target
was *actually* different than the destination (eg if the parent of a
straddled join wasn't already materialized, eg because it supported
query_through). This fixes that to use the target index rather than the
source index.

[0]: http://docs/dataflow/replay_paths.html#straddled-joins

Change-Id: I51d965594bb587caadeeb3b615bf4f467511098a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5041
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>