benchmarks: Add basic implementation of graphing results
Add a very basic implementation to benchmarks of running the same
benchmark multiple times but varying one of its parameters, and then
outputting the results to a CSV file. The idea here is that we can use
this to graph the output of a benchmark as any arbitrary input parameter
changes. Under the hood, this is (maybe somewhat hackily?) building a
set of *partial command line arguments* for the benchmark command, then
using `clap::Parser::update_from` to provide those command line
arguments to the benchmark command struct.

Currently there's nothing here to actually generate graphs, but I've
tested this with the `QueryBenchmark` varying `target-qps` like so:

cargo run --bin benchmarks -- \
  --benchmark ../src/yaml/benchmarks/minimal/point_query_small.yaml \
  --database-type mysql \
  --local \
  --graph \
  --x-axis target-qps \
  --x-values 1,10 \
  --graph-results-path ./results.csv

and we get a CSV with some data!

Change-Id: I1a756849a7457a8550c968e0832070acb679417c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5083
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI