helm: Initial pass at chart tests
Create two tests using terratest to test integration of the new Helm
chart. Tests are simple to start, only validating that we can add our
repo and render the charts template with correct version and namespace.
These tests will expand over time to cover much more of the stack as
well as perform e2e tests of the services themselves.

Feel free to use "go test" or bazel to run the tests.

- Remove WORKSPACE at root of repository to simplify bazel integration
  across different projects.
- Add .helmignore with entries to ignore the chart testing code and
  other non-chart files
- Set chart version to 0.4.0-rc3
- Add .gitignore files for `helm/readyset/` and `helm/charts/`
- Add plumbing for building with Bazel
- Cleanup some older bazel files from golang/
- Add TESTING.md for how to test locally

Change-Id: I567a6ae5b9cb6246b93f19e5c6d3861894a3cd2a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5073
Tested-by: Buildkite CI
Reviewed-by: Alana Marzoev <alana@readyset.io>