benchmarks/write_benchmark: Allow adding indices to tables
Support measuring the impact of adding indices to tables on write
latency with a new optional `--indices-per-table` flag in
write_benchmark. If this flag is passed, during setup the benchmark
will create that many indices in each table, by migrating a partial
query that is keyed on that many distinct permutations of the columns in
the table (if there aren't enough columns to support this we'll return
an error, but combinatorics is such that not a lot of columns are
necessary).

Currently, this only supports point indices - in the future we can add
support for range indices as well.

Change-Id: Ia5b1773bb63f2f69a40b02a87317ab361fc8f620
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5131
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>