system-benchmarks: fix wait_adapter_ready
The previous implementation of the `wait_adapter_ready` function didn't
work for Postgres for two reasons. First, it used
`DatabaseConnection::query` to query `SHOW READYSET STATUS`;
`DatabaseConnection::query` automatically prepares statements that
aren't already prepared, and since ReadySet can't prepare `SHOW`
queries, the query failed with an error. Second, the Postgres wire
protocol requires a database name when initiating a connection, so
we always need to include the name in the ReadySet URL in
system-benchmarks.

This commit makes use of `DatabaseConnection::simple_query` to ensure
that we don't try to prepare the `SHOW READYSET STATUS` query and ensures
that the ReadySet URL always ends with the database name.

Change-Id: I367e5e039a38b7e798b625bbde47ad81cab0de67
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5136
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>